refactor out initialisation of database and remove unessesary if condition
